Add files via upload
In response to the dynamic and competitive landscape of the food industry, our project set out to conduct a detailed analysis of pizza sales for restaurant with the goal of optimizing performance and increasing revenue. The primary objective was to understand customer preferences, identify areas for improvement, and implement strategies to enhance pizza sales.